+++ gensim/test/test_ldamodel.py
@@ -48,6 +48,12 @@ def testfile():
     return os.path.join(tempfile.gettempdir(), 'gensim_models.tst')
 
 
+def testRandomState():
+    testcases = [numpy.random.seed(0), None, numpy.random.RandomState(0), 0]
+    for testcase in testcases:
+        assert(isinstance(ldamodel.get_random_state(testcase), numpy.random.RandomState))
+        assertEqual(ldamodel.get_random_state(testcase), numpy.random.RandomState(0))
+
 class TestLdaModel(unittest.TestCase):
     def setUp(self):
         self.corpus = mmcorpus.MmCorpus(datapath('testcorpus.mm'))
@@ -251,8 +257,7 @@ class TestLdaModel(unittest.TestCase):
 
     def testGetDocumentTopics(self):
 
-        numpy.random.seed(0)
-        model = self.class_(self.corpus, id2word=dictionary, num_topics=2, passes= 100)
+        model = self.class_(self.corpus, id2word=dictionary, num_topics=2, passes= 100, random_state=numpy.random.seed(0))
 
         doc_topics = model.get_document_topics(self.corpus)
 
@@ -285,8 +290,7 @@ class TestLdaModel(unittest.TestCase):
 
     def testTermTopics(self):
 
-        numpy.random.seed(0)
-        model = self.class_(self.corpus, id2word=dictionary, num_topics=2, passes=100)
+        model = self.class_(self.corpus, id2word=dictionary, num_topics=2, passes=100, random_state=numpy.random.seed(0))
 
         # check with word_type
         result = model.get_term_topics(2)