+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes
   - NOT BACKWARDS COMPATIBLE!
 * Added random_state parameter to LdaState initializer and check_random_state() (@droudy, #113)
 * Implemented LsiModel.docs_processed attribute
+* `n_similarity()` raises `ValueError` if an empty list is passed to it in word2vec, doc2vec (@droudy, #761)
 
 0.13.1, 2016-06-22